+++ sbndcode/Calibration/TPCCalorimetry/NormalizeDriftSQLite_class.h
@@ -1,0 +1,4 @@
+#ifndef NORMALIZEDRIFTSQLITE_CLASS_H
+#define NORMALIZEDRIFTSQLITE_CLASS_H
+#include "sbndcode/Calibration/TPCCalorimetry/NormalizeDriftSQLite_Definitions.h"
+#endif
