+++ sbndcode/Calibration/TPCCalorimetry/NormalizeDriftSQLite_tool.h
@@ -1,0 +1,5 @@
+#ifndef NORMALIZEDRIFTSQLITE_TOOL_H
+#define NORMALIZEDRIFTSQLITE_TOOL_H
+#include "sbndcode/Calibration/TPCCalorimetry/NormalizeDriftSQLite_Definitions.h"
+DEFINE_ART_CLASS_TOOL(sbnd::calo::NormalizeDriftSQLite)
+#endif
